Refactor redundant Json response code
