+++ src/Panel/Dialog.php
@@ -13,22 +13,10 @@ namespace Kirby\Panel;
  * @copyright Bastian Allgeier GmbH
  * @license   https://getkirby.com/license
  */
-class Dialog
+class Dialog extends Json
 {
-    /**
-     * Renders the error dialog response with provided message
-     *
-     * @param string $message
-     * @param int $code
-     * @return array
-     */
-    public static function error(string $message, int $code = 404)
-    {
-        return [
-            'code'  => $code,
-            'error' => $message
-        ];
-    }
+
+    protected static $key = '$dialog';
 
     /**
      * Renders dialogs
@@ -39,30 +27,13 @@ class Dialog
      */
     public static function response($data, array $options = [])
     {
-        // handle Kirby exceptions
-        if (is_a($data, 'Kirby\Exception\Exception') === true) {
-            $data = static::error($data->getMessage(), $data->getHttpCode());
-
-        // handle exceptions
-        } elseif (is_a($data, 'Throwable') === true) {
-            $data = static::error($data->getMessage(), 500);
-
         // interpret true as success
-        } elseif ($data === true) {
+        if ($data === true) {
             $data = [
                 'code' => 200
             ];
-
-        // only expect arrays from here on
-        } elseif (is_array($data) === false) {
-            $data = static::error('Invalid dialog response', 500);
         }
 
-        // always inject the response code
-        $data['code']     = $data['code']    ?? 200;
-        $data['path']     = $options['path'] ?? null;
-        $data['referrer'] = Panel::referrer();
-
-        return Panel::json(['$dialog' => $data], $data['code']);
+        return parent::response($data, $options);
     }
 }
