+++ src/Panel/Dropdown.php
@@ -13,22 +13,10 @@ namespace Kirby\Panel;
  * @copyright Bastian Allgeier GmbH
  * @license   https://getkirby.com/license
  */
-class Dropdown
+class Dropdown extends Json
 {
-    /**
-     * Renders the error dropdown response with provided message
-     *
-     * @param string $message
-     * @param int $code
-     * @return array
-     */
-    public static function error(string $message, int $code = 404)
-    {
-        return [
-            'code'  => $code,
-            'error' => $message
-        ];
-    }
+
+    protected static $key = '$dropdown';
 
     /**
      * Renders dropdowns
@@ -39,30 +27,12 @@ class Dropdown
      */
     public static function response($data, array $options = [])
     {
-        // handle Kirby exceptions
-        if (is_a($data, 'Kirby\Exception\Exception') === true) {
-            $data = static::error($data->getMessage(), $data->getHttpCode());
-
-        // handle exceptions
-        } elseif (is_a($data, 'Throwable') === true) {
-            $data = static::error($data->getMessage(), 500);
-
-        // only expect arrays from here on
-        } elseif (is_array($data) === false) {
-            $data = static::error('Invalid dropdown response', 500);
-
-        // options are passed as a simple array
-        } else {
+        if (is_array($data) === true) {
             $data = [
                 'options' => $data
             ];
         }
 
-        // add common stuff to the data array
-        $data['code']     = $data['code']    ?? 200;
-        $data['path']     = $options['path'] ?? null;
-        $data['referrer'] = Panel::referrer();
-
-        return Panel::json(['$dropdown' => $data], $data['code']);
+        return parent::response($data, $options);
     }
 }
